Memory leak checking now works.
Go read the horrible hacks required.
As our /proc/self/<fd> are not symlinks but actual links, we just use them as the filenames. I dont see a major issue in that. If the fd is closed, whatever we read from it should be invalidated anyways.
Other assorted misery is also included in this commit


git-svn-id: http://svn.berlios.de/svnroot/repos/vg4nbsd/trunk@272 553a31c3-2ffc-0310-903e-cdb4b4d06021